+++ src/parsers/custom/problem/EOlympProblemParser.ts
@@ -1,8 +1,8 @@
-import * as $ from 'jquery';
 import { Parser } from '../../Parser';
 import { Sendable } from '../../../models/Sendable';
 import { Test } from '../../../models/Test';
 import { CustomTask } from '../../../models/CustomTask';
+import { htmlToElement } from '../../../utils';
 
 export class EOlympProblemParser extends Parser {
   getMatchPatterns(): string[] {
@@ -11,28 +11,28 @@ export class EOlympProblemParser extends Parser {
 
   parse(html: string): Promise<Sendable> {
     return new Promise(resolve => {
-      const $html = $(html);
+      const elem = htmlToElement(html);
 
-      const taskName = $html.find('.eo-paper__header').text();
+      const taskName = elem.querySelector('.eo-paper__header').textContent;
 
-      const contestNameParts = ['E-Olymp', $html.find('.eo-title__header').eq(1).text()];
-
+      const contestNameParts = ['E-Olymp', elem.querySelector('.eo-title__header').textContent];
       if (contestNameParts[1] === taskName) {
         contestNameParts.pop();
       }
 
       const contestName = contestNameParts.join(' - ');
 
-      const memoryLimit = parseInt($html.find('.eo-message__text b').last().text());
+      const memoryLimit = parseInt(elem.querySelectorAll('.eo-message__text b')[1].textContent);
 
       const tests: Test[] = [];
 
-      $html.find('.mdl-grid').has('.eo-code').each(function () {
-        const input = $(this).find('.eo-code').eq(0).text().trim();
-        const output = $(this).find('.eo-code').eq(1).text().trim();
+      const blocks = elem.querySelectorAll('.mdl-grid .eo-code');
+      for (let i = 0; i < blocks.length; i += 2) {
+        const input = blocks[i].textContent;
+        const output = blocks[i + 1].textContent;
 
         tests.push(new Test(input, output));
-      });
+      }
 
       resolve(new CustomTask(taskName, contestName, tests, memoryLimit));
     });