+++ src/parsers/default/contest/OldGoogleCodeJamContestParser.ts
@@ -1,8 +1,8 @@
 import { DefaultParser } from '../DefaultParser';
 import { DefaultWebsite } from '../../../models/DefaultTask';
 
-export class GoogleCodeJamOldContestParser extends DefaultParser {
-  website: DefaultWebsite = DefaultWebsite.GoogleCodeJam;
+export class OldGoogleCodeJamContestParser extends DefaultParser {
+  website: DefaultWebsite = DefaultWebsite.OldGoogleCodeJam;
 
   getMatchPatterns(): string[] {
     return ['https://code.google.com/codejam/contest/*/dashboard'];