+++ src/parsers/default/problem/NewGoogleCodeJamProblemParser.ts
@@ -1,0 +1,13 @@
+import { DefaultParser } from '../DefaultParser';
+import { DefaultWebsite } from '../../../models/DefaultTask';
+
+export class NewGoogleCodeJamProblemParser extends DefaultParser {
+  website: DefaultWebsite = DefaultWebsite.NewGoogleCodeJam;
+
+  getMatchPatterns(): string[] {
+    return [
+      'https://codejam.withgoogle.com/*/challenges/*/dashboard',
+      'https://codejam.withgoogle.com/*/challenges/*/dashboard/*',
+    ];
+  }
+}