+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.2.0 - Unknown
+- Fixed USACO parser
+- Added USACO Training parser
+- Removed jQuery dependency
+- Reverted CodeForces parser back to the CHelper one
+- Switched the new Google Code Jam parser to the CHelper one
+
 # 1.1.0 - April 1st, 2018
 - Added support for the new Google Code Jam website
 - Updated all dependencies